+++ src/app/orders/orders-routing.module.ts
@@ -2,11 +2,16 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 import { OrderComponent } from './components/order/order.component';
+import { NewOrderComponent } from './components/new-order/new-order.component';
 
 const routes: Routes = [
   {
     path: 'order',
     component: OrderComponent
+  },
+  {
+    path: 'new-order',
+    component: NewOrderComponent
   },
   {
     path: '',